+++ lib/bloom_remit_client/client.rb
@@ -65,6 +65,23 @@ module BloomRemitClient
       Responses::Payments::Create.new(raw_response: raw_response)
     end
 
+    def create_recipient(recipient_params={})
+      params = {recipient: recipient_params}.
+        merge(sender_id: recipient_params[:sender_id])
+      request_params = params.merge(default_params).merge(access_params)
+      raw_response = Requests::Recipients::Create.new(request_params).()
+      Responses::Recipients::Create.new(raw_response: raw_response)
+    end
+
+    def create_remittance(remittance_params={})
+      params = {remittance: remittance_params}.
+        merge(sender_id: remittance_params[:sender_id]).
+        merge(recipient_id: remittance_params[:recipient_id])
+      request_params = params.merge(default_params).merge(access_params)
+      raw_response = Requests::Remittances::Create.new(request_params).()
+      Responses::Remittances::Create.new(raw_response: raw_response)
+    end
+
     private
 
     # Should overwrite any other `:api_token`, `:api_secret`